add CardHelpers class for shared logic
